refactor: cli module to command submodules
