+++ src/scaffold/example.rs
@@ -5,28 +5,28 @@ use std::str::FromStr;
 
 #[derive(Serialize, Debug, Clone)]
 #[serde(tag = "type")]
-pub enum Example {
+pub enum ExampleType {
     HelloWorld,
     Forum,
 }
 
-impl std::fmt::Display for Example {
+impl std::fmt::Display for ExampleType {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         let str = match self {
-            Example::HelloWorld => "hello-world",
-            Example::Forum => "forum",
+            ExampleType::HelloWorld => "hello-world",
+            ExampleType::Forum => "forum",
         };
         write!(f, "{str}")
     }
 }
 
-impl FromStr for Example {
+impl FromStr for ExampleType {
     type Err = ScaffoldError;
 
     fn from_str(s: &str) -> ScaffoldResult<Self> {
         match s {
-            "hello-world" => Ok(Example::HelloWorld),
-            "forum" => Ok(Example::Forum),
+            "hello-world" => Ok(ExampleType::HelloWorld),
+            "forum" => Ok(ExampleType::Forum),
             _ => Err(ScaffoldError::InvalidExampleType(
                 s.to_string(),
                 "hello-world, forum".to_string(),
@@ -35,16 +35,16 @@ impl FromStr for Example {
     }
 }
 
-pub fn choose_example() -> ScaffoldResult<Example> {
+pub fn choose_example() -> ScaffoldResult<ExampleType> {
     let selection = Select::with_theme(&ColorfulTheme::default())
         .with_prompt("Choose example:")
-        .item(Example::HelloWorld)
-        .item(Example::Forum)
+        .item(ExampleType::HelloWorld)
+        .item(ExampleType::Forum)
         .default(0)
         .interact()?;
 
     match selection {
-        0 => Ok(Example::HelloWorld),
-        _ => Ok(Example::Forum),
+        0 => Ok(ExampleType::HelloWorld),
+        _ => Ok(ExampleType::Forum),
     }
 }
